requirements.txt
fastapi
uvicorn
python-multipart
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
oauth2client
requests
